+++ lizard_map/static/lizard_map/lizard_map.js
@@ -173,7 +173,45 @@ function setUpDateRangePopup() {
     });
 }
 
-function setWorkspaceSavePopup() {
+function setUpWorkspaceLoad() {
+    $(".workspace-load").live("click", function (event) {
+        var url;
+        event.preventDefault();
+        url = $(this).attr("href");
+        id = $(this).attr("data-workspace-id");
+        $.post(
+            url,
+            {
+                id: id
+            },
+            function(data) {
+            	result = $.parseJSON(data);
+            	if (result['redirect'])
+            	    window.location.href = result['redirect'];
+            }
+        );
+    });
+}
+
+function workspaceSavePopup(data) {
+    show_popup(data);
+    $('#workspace-save-submit').click(function(event) {
+        event.preventDefault();
+        $form = $('#workspace-save-form');
+        $.post(
+            $form.attr("action"), $form.serialize()
+        )
+        .success(
+            function (data) {
+                // send result to popup
+                show_popup(data);
+            }
+        );
+        return false;
+    });
+}
+
+function setUpWorkspaceSavePopup() {
     $(".popup-workspace-save").live("click", function (event) {
         var url;
         event.preventDefault();
@@ -181,7 +219,7 @@ function setWorkspaceSavePopup() {
         $.get(
             url,
             function(data) {
-                show_popup(data);
+                workspaceSavePopup(data);
             }
         );
     });
@@ -1168,7 +1206,8 @@ $(document).ready(function () {
     setUpLegendEdit();
 
     setUpMapLoadDefaultLocation();
-    setWorkspaceSavePopup();
+    setUpWorkspaceLoad();
+    setUpWorkspaceSavePopup();
     setUpDateRangePopup();
     setUpCollageEditPopup();
     setUpCollageTablePopup();